tools/crawler/requirements.txt
requests
beautifulsoup4
google-genai